requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
angelcommunity==2.0.0
annotated-types==0.7.0
anyio==4.11.0
ASLPAw==2.2.0
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
backrefs==5.9
bayanpy==0.7.7
bimlpa==0.1.2
black==25.9.0
cdlib==0.4.0
certifi==2025.8.3
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
complete-shuffle==1.1.2
contourpy==1.2.1
count-dict==1.1.1
cryptography==46.0.1
cycler==0.12.1
databind==4.5.2
databind.core==4.5.2
databind.json==4.5.2
decorator==5.2.1
demon==2.0.6
Deprecated==1.2.18
distro==1.9.0
docspec==2.2.1
docspec-python==2.2.2
docstring_parser==0.11
dynetx==0.3.2
eva-lcd==0.1.1
fastapi==0.117.1
fastuuid==0.12.0
filelock==3.19.1
fonttools==4.53.1
frozenlist==1.7.0
fsspec==2025.9.0
future==1.0.0
ghp-import==2.1.0
gmpy2==2.2.1
grpcio==1.75.0
gurobipy==12.0.3
h11==0.16.0
h2==4.3.0
hf-xet==1.1.10
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.1
hyperframe==6.1.0
idna==3.10
igraph==0.11.9
importlib_metadata==8.7.0
infomap==2.8.0
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.5.2
leidenalg==0.10.2
Levenshtein==0.27.1
litellm==1.77.4
loguru==0.7.3
madoka==0.7.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
mem0ai==0.1.117
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.6.20
mkdocs-material-extensions==1.3.1
multidict==6.6.4
multivalued-dict==2.0.1
munkres==1.1.4
mypy_extensions==1.1.0
narwhals==2.5.0
neo4j==5.28.2
networkx==3.5
nr-date==2.1.0
nr-stream==1.1.5
nr.util==0.8.12
numpy
openai==1.99.9
packaging
paginate==0.5.7
pandas==2.3.2
pathspec==0.12.1
pillow
platformdirs==4.4.0
plotly==6.3.0
pondpond==1.4.1
pooch==1.8.2
portalocker==3.2.0
posthog==6.7.6
propcache==0.3.2
protobuf==5.29.5
psutil==7.1.0
psycopg2-binary==2.9.10
PuLP==3.3.0
pure-nrng==1.1.0
pure-prng==2.9.0
pycairo==1.28.0
pyclustering==0.10.1.2
pycombo==0.1.8
pycparser==2.22
pydantic==2.11.9
pydantic_core==2.33.2
pydoc-markdown==4.8.2
Pygments==2.19.2
PyGObject==3.50.0
pymdown-extensions==10.16.1
pymilvus==2.6.2
PyMuPDF==1.26.4
pyparsing==3.2.0
pypdf==6.1.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-igraph==0.11.9
python-Levenshtein==0.27.1
python-louvain==0.16
pytokens==0.1.10
pytz==2025.2
PyYAML==6.0.2
pyyaml_env_tag==1.1
qdrant-client==1.15.1
randomgen==2.1.1
RapidFuzz==3.14.1
redis==6.4.0
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
rich==14.1.0
rpds-py==0.27.1
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
setuptools==78.1.1
shellingham==1.5.4
shuffle-graph==2.2.0
similarity-index-of-label-graph==2.0.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
texttable==1.7.0
threadpoolctl==3.6.0
thresholdclustering==1.1
tiktoken==0.11.0
tokenizers==0.22.1
tomli==2.2.1
tomli_w==1.2.0
tqdm==4.67.1
typeapi==2.2.4
typer==0.19.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
watchdog==6.0.0
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.3
wurlitzer==3.1.1
yapf==0.43.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.25.0